input: -i vmx: Add support for NVMe devices
We model NVMe devices in the source hypervisor.

We currently assume that no one is using the namespaces feature of
NVMe, ie. that each source device will appear in a Linux guest as
/dev/nvme0n1, /dev/nvme1n1, etc.  We could fix this if it is a
problem, but it requires adjusting the current assumption for
removable devices that slots are simple integers.

The devices are mapped to virtio-blk, so in the target the device name
has to change from /dev/nvme0 to /dev/vda (etc.)

Reported-by: Ming Xie
Fixes: https://bugzilla.redhat.com/show_bug.cgi?id=2070530
Reviewed-by: Laszlo Ersek <lersek@redhat.com>
(cherry picked from commit 75872bf282d7f2322110caca70963717b43806b1)